rimozione di import superflui,cambio del  meccanismo di gestione degli ordini,creazione di una pagina per lo storico ordini e di una per i dettagli degli ordini,cambi estetici alla pagina per inserire un metodo di pagamento ,implementazione funzionante dello stub di product control

+++ TCGMarket/src/main/java/model/ItemOrder.java
@@ -15,47 +15,47 @@ package model;
 
 public class ItemOrder {
 
-	 private ProductBean item;
-	  private int numItems;
-	  public ItemOrder(ProductBean item) {
-	    setItem(item);
-	    setNumItems(1);
-	  }
-	  public ProductBean getItem() {
-	    return(item);
-	  }
-	  protected void setItem(ProductBean item) {
-	    this.item = item;
-	  }
-	  public int getId() {
-	    return(getItem().getId());
-	  }
-	  public String getDescrizione() {
-	    return(getItem().getDescrizione());
-	  }
-	  
-	  public String getNome() {
-		  return(getItem().getNome());
-	  }
-	 
-	  public double getPrezzo() {
-		  return(Math.round(getItem().getPrezzo()*100.0))/100.0;
-	  }
-	  
-	  public int getNumItems() {
-	    return(numItems);
-	  }
-	  public void setNumItems(int n) {
-	    this.numItems = n;
-	  }
-	  public void incrementNumItems() {
-	    setNumItems(getNumItems() + 1);
-	  }
-	  public void cancelOrder() {
-	    setNumItems(0);
-	  }
-	  public double getTotalCost() {
-		  return (Math.round(getNumItems() * getPrezzo()*100)/100.00);
-	  }
+	private ProductBean item;
+	private int numItems;
+	public ItemOrder(ProductBean item) {
+		setItem(item);
+		setNumItems(1);
+	}
+	public ProductBean getItem() {
+		return(item);
+	}
+	protected void setItem(ProductBean item) {
+		this.item = item;
+	}
+	public int getId() {
+		return(getItem().getId());
+	}
+	public String getDescrizione() {
+		return(getItem().getDescrizione());
+	}
+
+	public String getNome() {
+		return(getItem().getNome());
+	}
+
+	public float getPrezzo() {
+		return (float) ((Math.round(getItem().getPrezzo()*100.0))/100.0);
+	}
+
+	public int getNumItems() {
+		return(numItems);
+	}
+	public void setNumItems(int n) {
+		this.numItems = n;
+	}
+	public void incrementNumItems() {
+		setNumItems(getNumItems() + 1);
+	}
+	public void cancelOrder() {
+		setNumItems(0);
+	}
+	public float getTotalCost() {
+		return (float) (Math.round(getNumItems() * getPrezzo()*100)/100.00);
+	}
 
 }
